requirements.txt
certifi>=2020
chardet==3.0.2
decorator==4.0.2
dlib==19.21.1
docx==0.2.4
idna==2.5
imageio==2.5
imageio-ffmpeg==0.2.0
librosa==0.8.0
moviepy==1.0.3
moviepy==1.0.3
networkx==2.5
nltk==3.5
numpy>=1.20
opencv-python>=4.5.1
pandas==1.2.3
Pillow==8.1.2
pip==21.0.1
proglog>=0.1
PyAudio==0.2.11
requests==2.8.1
SpeechRecognition==3.8.1
tensorflow==2.4
#tkinter==8.5
tqdm==4.11.2
urllib3!=1.25.0,!=1.25.1==2.8.1
Wave==0.0.2

# emailAPI requirements:
Flask==1.1.2
Flask-Mail==0.9.1
python-decouple==3.4

